+++ packages/react-components/react-card/library/src/CardPreview.ts
@@ -1,8 +1,15 @@
-export type { CardPreviewProps, CardPreviewSlots, CardPreviewState } from './components/CardPreview/index';
+export type {
+  CardPreviewProps,
+  CardPreviewSlots,
+  CardPreviewState,
+  CardPreviewBaseProps,
+  CardPreviewBaseState,
+} from './components/CardPreview/index';
 export {
   CardPreview,
   cardPreviewClassNames,
   renderCardPreview_unstable,
   useCardPreviewStyles_unstable,
   useCardPreview_unstable,
+  useCardPreviewBase_unstable,
 } from './components/CardPreview/index';