+++ packages/react-components/react-radio/library/src/components/Radio/index.ts
@@ -1,5 +1,12 @@
 export { Radio } from './Radio';
-export type { RadioOnChangeData, RadioProps, RadioSlots, RadioState } from './Radio.types';
+export type {
+  RadioOnChangeData,
+  RadioProps,
+  RadioSlots,
+  RadioState,
+  RadioBaseProps,
+  RadioBaseState,
+} from './Radio.types';
 export { renderRadio_unstable } from './renderRadio';
-export { useRadio_unstable } from './useRadio';
+export { useRadio_unstable, useRadioBase_unstable } from './useRadio';
 export { radioClassNames, useRadioStyles_unstable } from './useRadioStyles.styles';