+++ packages/react-components/react-radio/library/src/components/Radio/useRadio.tsx
@@ -5,7 +5,7 @@ import { Label } from '@fluentui/react-label';
 import { getPartitionedNativeProps, mergeCallbacks, useId, slot } from '@fluentui/react-utilities';
 import { useRadioGroupContextValue_unstable } from '../../contexts/RadioGroupContext';
 import { useFocusWithin } from '@fluentui/react-tabster';
-import type { RadioProps, RadioState } from './Radio.types';
+import type { RadioBaseProps, RadioBaseState, RadioProps, RadioState } from './Radio.types';
 
 /**
  * Create the state required to render Radio.
@@ -17,6 +17,29 @@ import type { RadioProps, RadioState } from './Radio.types';
  * @param ref - reference to `<input>` element of Radio
  */
 export const useRadio_unstable = (props: RadioProps, ref: React.Ref<HTMLInputElement>): RadioState => {
+  const state = useRadioBase_unstable(props, ref);
+
+  return {
+    ...state,
+    // eslint-disable-next-line @typescript-eslint/no-deprecated
+    components: { ...state.components, label: Label },
+    label: slot.optional(props.label, {
+      defaultProps: { ...state.label },
+      elementType: Label,
+    }),
+  };
+};
+
+/**
+ * Create the state required to render Radio.
+ *
+ * The returned state can be modified with hooks such as useRadioStyles_unstable,
+ * before being passed to renderRadio_unstable.
+ *
+ * @param props - props from this instance of Radio
+ * @param ref - reference to `<input>` element of Radio
+ */
+export const useRadioBase_unstable = (props: RadioBaseProps, ref: React.Ref<HTMLInputElement>): RadioBaseState => {
   const group = useRadioGroupContextValue_unstable();
 
   const {
@@ -61,7 +84,7 @@ export const useRadio_unstable = (props: RadioProps, ref: React.Ref<HTMLInputEle
   input.onChange = mergeCallbacks(input.onChange, ev => onChange?.(ev, { value: ev.currentTarget.value }));
   const label = slot.optional(props.label, {
     defaultProps: { htmlFor: input.id, disabled: input.disabled },
-    elementType: Label,
+    elementType: 'label',
   });
   const indicator = slot.always(props.indicator, {
     defaultProps: { 'aria-hidden': true },
@@ -69,7 +92,7 @@ export const useRadio_unstable = (props: RadioProps, ref: React.Ref<HTMLInputEle
   });
   return {
     labelPosition,
-    components: { root: 'span', input: 'input', label: Label, indicator: 'div' },
+    components: { root: 'span', input: 'input', label: 'label', indicator: 'div' },
     root,
     input,
     label,