+++ packages/react-components/react-card/library/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui/react-card
 
-This log was last generated on Wed, 25 Feb 2026 13:32:28 GMT and should not be manually modified.
+This log was last generated on Thu, 26 Mar 2026 08:12:48 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [9.6.0](https://github.com/microsoft/fluentui/tree/@fluentui/react-card_v9.6.0)
+
+Thu, 26 Mar 2026 08:12:48 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-card_v9.5.11..@fluentui/react-card_v9.6.0)
+
+### Minor changes
+
+- feat: add base hooks for Card ([PR #35824](https://github.com/microsoft/fluentui/pull/35824) by dmytrokirpa@microsoft.com)
 
 ## [9.5.11](https://github.com/microsoft/fluentui/tree/@fluentui/react-card_v9.5.11)